checkpoint(CR-0001): gaps fixed per validation report
- NFR-1: split rasterisation helpers into export-png.rasterize.ts sibling
  per the CR's documented escape hatch; export-png.ts now 236 lines (<300)
- NFR-3: documented as manual-verification-required (CR delegates timing
  to local review on 2024-class Chromium)
- AC-8 / NFR-5: reconcile wording to "no NEW warnings introduced"; the
  two pre-existing biome warnings are outside this CR's Affected Components

All 5 specified tests still pass. Public API unchanged.
